notify_service/requirements.txt
pyinotify
aiohttp